[None][perf] ltx2: kernel accepts bf16 cos/sin, drop .float() upcast in helper
Add a bf16 branch to the fused full-dim RMSNorm+RoPE kernels so the
helper no longer has to upcast cos/sin from bf16 to fp32 before the
launch. Two kernels gain a `typename CosT` template parameter:
  - fusedDiTSplitQKNormRopeKernel    (LTX-2 cross-attn / SEPARATE_QKV)
  - fusedDiTQKNormFullDimRopeKernel  (LTX-2 self-attn / FUSE_QKV)

Both keep the existing fp32 cos path; bf16 cos is loaded as a single
LDG.128 (uint4) per chunk vs 2x LDG.128 (float4) for fp32, and upcast
in registers via __bfloat1622float2 before the math (which still runs
in fp32). Math is unchanged: bf16 -> fp32 is lossless, so kernel output
is bit-equivalent to the fp32-cos path.

Op bindings (`fusedDiTSplitQKNormRopeOp.cpp`, `fusedDiTQKNormRopeOp.cpp`)
relax the cos/sin dtype check from strict-fp32 to fp32-or-bf16 and
forward `cos_is_bf16` to the launcher. The per-head FLUX/Cosmos path
(`fusedDiTQKNormRopeKernel`, non-FullDim) keeps fp32-only since FLUX
cos is already fp32 upstream.

Helpers `LTX2Attention._apply_split_norm_rope` and
`Attention.apply_qk_norm_rope` drop `.float()` from the cos chain;
cos.reshape(..).contiguous() now passes bf16 through directly. For
LTX-2 this skips the per-call bf16->fp32 alloc + cast (~200 -> 400 MB
on video shape). For FLUX this is a no-op since cos was already fp32.

Unit tests parametrize on cos_dtype in {fp32, bf16}; 104/104 pass on
B200 in 1.28s, including LTX-2 production shapes
(T=12288 H=32 D=128, T=504 H=32 D=64) for both interleave modes.

E2E: 40-step LTX-2 single-stage produces a bit-identical video
(sha256) vs the prior B-1 commit. Single-op kernel time on B200
(B=2, 20-iter median) drops on LTX-2 cells:
  ltx2_video_self_attn   (T=12288 H=32 D=128)  0.3029 -> 0.2464 ms (-19%)
  ltx2_audio_self_attn   (T=504   H=32 D=64)   0.0095 -> 0.0086 ms (-10%)
  ltx2_text_cross_video  (T=12288 H=32 D=128)  0.1789 -> 0.1387 ms (-22%)
  ltx2_av_a2v_q          (T=12288 H=32 D=128)  0.1789 -> 0.1386 ms (-22%)

FLUX per-head cells unchanged within noise.

+++ tensorrt_llm/_torch/visual_gen/models/ltx2/transformer_ltx2.py
@@ -234,8 +234,9 @@ class LTX2Attention(Attention):
             )
         # cos/sin are token-major [B, T, H, D] (or [B, T, D] for shared per-token);
         # reshape(-1, cos_last) yields the [B*T, cos_last] layout the kernel reads.
-        cos_2d = cos.reshape(-1, cos_last).float().contiguous()
-        sin_2d = sin.reshape(-1, cos_last).float().contiguous()
+        # B-2: kernel accepts bf16 cos (upcasts in registers); skip the .float() cast.
+        cos_2d = cos.reshape(-1, cos_last).contiguous()
+        sin_2d = sin.reshape(-1, cos_last).contiguous()
         if cos_2d.shape[0] == T and B > 1:
             cos_tiled = cos_2d.repeat(B, 1)
             sin_tiled = sin_2d.repeat(B, 1)
